+++ compute/worker/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "noEmit": true
+    "noEmit": true,
   },
   "include": ["src/**/*.ts"],
   "exclude": []
